modified compile settings and respected architecture
